+++ lib/classes/hive_type_values.dart
@@ -4,4 +4,5 @@ class HiveTypesEnum {
   static const int config = 2;
   static const int configGraph = 3;
   static const int configLog = 4;
+  static const int configNotify = 5;
 }